add button to dropdown, rename to dummy dropdown.
